+++ telegram-bot/src/main/kotlin/com/ajaxproject/telegrambot/bot/handlers/impl/financehandlers/GetCurrentBalanceHandler.kt
@@ -42,7 +42,6 @@ class GetCurrentBalanceHandler(
                     dispatchRequest.updateSession.localization
                 )
             }
-            .then(telegramService.returnToMainMenu(dispatchRequest))
             .thenReturn(Unit)
     }
 }
