+++ telegram-bot/src/main/resources/texts/texts-en.properties
@@ -12,7 +12,8 @@ CURRENCY_TEXT = Select the currency:
 FINANCE_MANAGE_TEXT = Select an option:
 ADD_FINANCE_TEXT = Enter the amount and the description you want to add in the following form.\nExample:\nFor incomes: +1000 Salary\nFor expenses: -1000 Rent
 SUCCESSFUL_ADD_FINANCE_TEXT = Your data have successfully added.\nIf you want you can add data again or return to the main menu
-FAILED_ADD_FINANCE_TEXT = Your data is not added, please fill it in according to the example
+FAILED_ADD_FINANCE_TEXT = Your data is not added, please fill it again in according to the example or return to the main menu
+NO_FINANCE_TEXT = You have no finances yet
 
 #BUTTONS
 GET_CURRENCY_USD_BUTTON = Get USD rate
@@ -25,5 +26,5 @@ GET_ALL_EXPENSES_BUTTON = Get all expenses
 GET_CURRENCY_RATE_BUTTON = Get currency rate
 MANAGE_FINANCES_MENU_BUTTON = Manage finances
 GET_CURRENT_BALANCE_BUTTON = Get current balance
-BACK_TO_MENU_BUTTON_BUTTON = Main menu
-DELETE_DATA_BUTTON_BUTTON = Delete your data
+BACK_TO_MENU_BUTTON = Main menu
+DELETE_DATA_BUTTON = Delete your data
