+++ telegram-bot/src/main/resources/texts/texts-ua.properties
@@ -12,7 +12,8 @@ CURRENCY_TEXT = Виберіть валюту:
 FINANCE_MANAGE_TEXT = Виберіть опцію:
 ADD_FINANCE_TEXT = Введіть суму та опис, яку ви хочете додати в наступній формі.\nПриклад:\nДля доходів: +1000 Зарплата\nДля витрат: -1000 Квартплата
 SUCCESSFUL_ADD_FINANCE_TEXT = Ваші дані успішно додано.\nЯкщо ви бажаєте, ви можете додати дані ще раз або повернутися до головного меню
-FAILED_ADD_FINANCE_TEXT = Ваші дані не додані, будь ласка, заповніть їх відповідно до прикладу
+FAILED_ADD_FINANCE_TEXT = Ваші дані не додані, будь ласка, заповніть їх відповідно до прикладу знову або поверніться до головного меню
+NO_FINANCE_TEXT = Ви не додали жодних фінансів
 
 #BUTTONS
 GET_CURRENCY_USD_BUTTON = Отримати курс долара США
